Refactor lib into separate modules
