+++ lib/getNewUser.js
@@ -11,14 +11,4 @@ const getNewUser = async (id) => {
   return user;
 };
 
-const mapObjectToArray = (o, cb) => {
-  const results = [];
-
-  for (const [k, v] of Object.entries(o)) {
-    results.push(cb(k, v, o));
-  }
-
-  return results;
-};
-
-module.exports = { getNewUser, mapObjectToArray };
+module.exports = getNewUser;
