+++ lib/index.js
@@ -1,0 +1,4 @@
+const getNewUser = require("./getNewUser");
+const mapObjectToArray = require("./mapObjectToArray");
+
+module.exports = { getNewUser, mapObjectToArray };
